py/requirements.txt
fonttools
brotli
pillow
pillow-heif